requirements.txt
playwright.sync_api
pandas
openpyxl
termcolor
requests
uuid